+++ src/Logitar.Portal.Core/Configurations/Commands/InitializeConfigurationHandler.cs
@@ -1,4 +1,5 @@
 ﻿using AutoMapper;
+using Logitar.EventSourcing;
 using Logitar.Portal.Core.Realms;
 using Logitar.Portal.Core.Users;
 using Logitar.Portal.Core.Users.Contact;
@@ -9,17 +10,14 @@ namespace Logitar.Portal.Core.Configurations.Commands;
 
 internal class InitializeConfigurationHandler : IRequestHandler<InitializeConfiguration, Configuration>
 {
-  private readonly IApplicationContext _applicationContext;
   private readonly IConfigurationRepository _configurationRepository;
   private readonly IMapper _mapper;
   private readonly IUserRepository _userRepository;
 
-  public InitializeConfigurationHandler(IApplicationContext applicationContext,
-    IConfigurationRepository configurationRepository,
+  public InitializeConfigurationHandler(IConfigurationRepository configurationRepository,
     IMapper mapper,
     IUserRepository userRepository)
   {
-    _applicationContext = applicationContext;
     _configurationRepository = configurationRepository;
     _mapper = mapper;
     _userRepository = userRepository;
@@ -40,15 +38,15 @@ internal class InitializeConfigurationHandler : IRequestHandler<InitializeConfig
     ReadOnlyPasswordSettings? passwordSettings = ReadOnlyPasswordSettings.From(input.PasswordSettings);
     ReadOnlyLoggingSettings? loggingSettings = ReadOnlyLoggingSettings.From(input.LoggingSettings);
 
-    // TODO(fpion): user should create itself and the configuration
+    AggregateId userId = AggregateId.NewId();
 
-    ConfigurationAggregate configuration = new(_applicationContext.ActorId, defaultLocale, input.Secret,
+    ConfigurationAggregate configuration = new(userId, defaultLocale, input.Secret,
       usernameSettings, passwordSettings, loggingSettings);
 
-    UserAggregate user = new(_applicationContext.ActorId, configuration.UsernameSettings, userInput.Username, userInput.FirstName,
-      lastName: userInput.LastName, locale: defaultLocale);
-    user.ChangePassword(_applicationContext.ActorId, configuration.PasswordSettings, userInput.Password);
-    user.SetEmail(_applicationContext.ActorId, new ReadOnlyEmail(userInput.EmailAddress));
+    UserAggregate user = new(userId, configuration.UsernameSettings, userInput.Username, userInput.FirstName,
+      lastName: userInput.LastName, locale: defaultLocale, id: userId);
+    user.ChangePassword(user.Id, configuration.PasswordSettings, userInput.Password);
+    user.SetEmail(user.Id, new ReadOnlyEmail(userInput.EmailAddress));
 
     await _configurationRepository.SaveAsync(configuration, cancellationToken);
     await _userRepository.SaveAsync(user, cancellationToken);